backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6
diffusers==0.16.1
transformers==4.28.1
accelerate==0.19.0
safetensors==0.3.1
pillow==9.5.0
numpy==1.24.3
torch>=2.0.0
psutil==5.9.5
pyjwt==2.6.0